feat(ops): split ecdsa by curve
