+++ crates/accelerators/src/ops/ecdsa/secp256k1.rs
@@ -1,5 +1,8 @@
-//! ECDSA operations.
+//! ECDSA over the secp256k1 curve.
 
+// In the guest, secp256k1 operations use the OpenVM-accelerated k256; on
+// the host they use upstream RustCrypto k256 (the ECDSA recovery
+// relies on zkVM hints and is unimplemented outside the guest).
 #[cfg(any(target_os = "none", target_os = "openvm"))]
 use openvm_k256 as k256;
 
@@ -7,10 +10,7 @@ use k256::ecdsa::{signature::hazmat::PrehashVerifier, RecoveryId, Signature, Ver
 
 use crate::{
     ops::Error,
-    types::{
-        ZkvmSecp256k1Hash, ZkvmSecp256k1Pubkey, ZkvmSecp256k1Signature, ZkvmSecp256r1Hash,
-        ZkvmSecp256r1Pubkey, ZkvmSecp256r1Signature,
-    },
+    types::{ZkvmSecp256k1Hash, ZkvmSecp256k1Pubkey, ZkvmSecp256k1Signature},
 };
 
 /// Recover the uncompressed secp256k1 public key from an ECDSA signature
@@ -66,29 +66,6 @@ pub fn secp256k1_verify(
     if let Some(normalized) = signature.normalize_s() {
         signature = normalized;
     }
-    *verified = key.verify_prehash(&msg.data, &signature).is_ok();
-    Ok(())
-}
-
-/// Verify an ECDSA signature over secp256r1 (P-256) against an uncompressed
-/// public key, writing the result to `verified`.
-pub fn secp256r1_verify(
-    msg: &ZkvmSecp256r1Hash,
-    sig: &ZkvmSecp256r1Signature,
-    pubkey: &ZkvmSecp256r1Pubkey,
-    verified: &mut bool,
-) -> Result<(), Error> {
-    use openvm_p256::{
-        ecdsa::{Signature, VerifyingKey},
-        EncodedPoint,
-    };
-
-    *verified = false;
-
-    let encoded_point = EncodedPoint::from_untagged_bytes(&pubkey.data.into());
-    let key =
-        VerifyingKey::from_encoded_point(&encoded_point).map_err(|_| Error::PointNotOnCurve)?;
-    let signature = Signature::from_slice(&sig.data).map_err(|_| Error::InvalidSignature)?;
     *verified = key.verify_prehash(&msg.data, &signature).is_ok();
     Ok(())
 }
